Responding to review and fixing package issue with rocky
