+++ wake.spec.in
@@ -32,6 +32,7 @@ mkdir -p %{buildroot}/var/cache/wake
 %files
 /usr/bin/wake
 /usr/bin/wakebox
+/usr/bin/job-cache
 /usr/lib/wake
 /usr/share/wake
 %dir /var/cache/wake
